// lib/api/generated/schemas/generateImagesDtoStyle.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * НеМалика API
 * Маркетплейс компьютерной техники — REST API для веб-витрины (Next.js) и Telegram mini-app. Покупать можно без авторизации; вход через Telegram нужен продавцу, администратору и покупателю, который хочет видеть историю просмотров на всех своих устройствах.
 * OpenAPI spec version: 1.0.0
 */

export type GenerateImagesDtoStyle = typeof GenerateImagesDtoStyle[keyof typeof GenerateImagesDtoStyle];


export const GenerateImagesDtoStyle = {
  infographic: 'infographic',
  photo: 'photo',
} as const;